게시글 빈칸일 경우 작성불가 기능 추가

+++ src/main/java/com/springweek2/springweek2/service/MemoService.java
@@ -37,7 +37,7 @@ public class MemoService {
             memoRepository.deleteById(id);
             return id;
         }
-        return (long)-1;
+        return -1L;
     }
 
     @Transactional
@@ -53,7 +53,7 @@ public class MemoService {
             memoRepository.save(memo);
             return id;
         }
-        return (long)-1;
+        return -1L;
     }
 
     public Memo ModalMemo(Long id) {
